# src/main/resources/api-docs.yaml
openapi: 3.1.0
info:
  title: Boosting platform
  version: "1.0"
servers:
  - url: http://localhost:6969
    description: Generated server url
security:
  - BearerAuth: []
...
paths:
  /api/users/me/nickname:
    post:
      tags:
        - Профили пользователей
      summary: Изменить никнейм
      operationId: changeNickname
      parameters:
        - name: nickname
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /api/users/me/description:
    post:
      tags:
        - Профили пользователей
      summary: Изменить описание профиля
      operationId: changeDescriptionProfile
      parameters:
        - name: description
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
...
  /api/boosters/finance/withdrawals:
    post:
      tags:
        - Финансы бустеров
      summary: Запросить вывод средств
      operationId: handleWithdrawal
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HandleWithdrawalRqDto"
        required: true
      responses:
        "200":
          description: OK
  /api/boosters/finance/tips:
    post:
      tags:
        - Финансы бустеров
      summary: Отправить чаевые
      operationId: sendTip
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HandleSendTipRqDto"
        required: true
      responses:
        "200":
          description: OK
...
  /api/public/offers/games/{gameId}:
    get:
      tags:
        - Публичные предложения
      summary: Получить предложения по ID игры
      operationId: getOffersByGameId
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameOffersRsDto"
...
  /api/games/{secondId}:
    get:
      tags:
        - Каталог игр
      summary: Получить игру с категориями по ID
      operationId: getGameWithCategories
      parameters:
        - name: secondId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GameBySecondIdRsDto"
...
  /api/boosters/finance/orders/{orderId}/tips:
    get:
      tags:
        - Финансы бустеров
      summary: Получить историю чаевых по заказу
      operationId: getOrderTipHistory
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderTipHistoryRsDto"
...
components:
  schemas:
...
    GameOffersRsDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор предложения
          example: 70b82203-0f41-47e1-b11c-0fb96f8f9204
        title:
          type: string
          description: Название предложения
          example: Rank boosting
        description:
          type: string
          description: Описание предложения
          example: Very long description
        imageUrl:
          type: string
          description: Ссылка на картинку предложения
          example: https://bb82cdcce70076216efdbdfb864e275d.jpg
        price:
          type: number
          description: Стоимость предложения
          example: 100
...
    HandleWithdrawalRqDto:
      type: object
      properties:
        withdrawalAmount:
          type: number
          description: Сумма для запроса на вывод средств бустера
          example: 99.99
    HandleSendTipRqDto:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
          description: Идентификатор заказа
          example: 123e4567-e89b-12d3-a456-426614174000
        tipAmount:
          type: number
          description: Сумма чаевых бустеру за заказ
          example: 9.99
      required:
        - orderId
        - tipAmount
...
    GameBySecondIdRsDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор игры
          example: 70b82203-0f41-47e1-b11c-0fb96f8f9204
        secondId:
          type: string
          description: Второй идентификатор игры
          example: ABCD-123
        name:
          type: string
          description: Название игры
          example: Destiny 2
        categories:
          type: array
          items:
            description: "Список категорий, к которым относится игра"
...
    OrderTipHistoryRsDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор записи
          example: 70b82203-0f41-47e1-b11c-0fb96f8f9204
        status:
          type: string
          description: Статус чаевых
          enum:
            - ON_PENDING
            - COMPLETED
          example: ON_PENDING
        amount:
          type: number
          description: Сумма чаевых
          example: 5.55
        createdAt:
          type: string
          format: date-time
          description: Дата и время (UTC) создания операции
          example: 2025-06-26 12:00
...
  securitySchemes:
    BearerAuth:
      type: http
      name: BearerAuth
      scheme: bearer
      bearerFormat: JWT